Add a method to return if a modifier is active
In case one would like to implement behaviour that changes when a
modifier is active, one would need to know the state of the modifier.
While it is possible to track this on a higher level, this is something
already available in the report. As such, the most straightforward way
to achieve this goal is to have a function that looks into the report.

This patch does just that.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>